third_party/WebKit/LayoutTests/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_dynamic_insertion-expected.txt

Description

This test validates that all of the initiator types are represented even when dynamically inserted.



PASS http://127.0.0.1:8000/w3c/webperf/resources/generate_resource.php?type=css is expected to have initiatorType link 
PASS http://127.0.0.1:8000/w3c/webperf/resources/generate_resource.php?type=iframe is expected to have initiatorType iframe 
PASS http://127.0.0.1:8000/w3c/webperf/resources/generate_resource.php?type=image is expected to have initiatorType img 
